Add support for Modular Event Bridge for Log Ingestion (single) (#4)
* Add support for Modular Event Bridge for Log Ingestion (single)

Change summary:
----------------
- Added the module as an integration module since it is feature
  agnostic and can be used for shared integrations.
- Added the respective tf files for single onboarding case
- Added test example for event-bridge integration (single)

Testing:
---------
Validated the changes with single onboarding of actual accounts on AWS.

* Add README and minor fixes

* Remove provider block in snippet

As suspected, the provider block for primary region is no longer needed in event bridge snippets. The foundational snippet takes care of it. Was able to test this end to end.

* Update README and remove delegated_admin check for single installs

* Fix ciem feature dependency

The feature resource should depend on the respective features
and not the modules. Fixing that. Tested this with a terraform
run with apply & destroy.